more complete code separation
